pyproject.toml
[project]
name = "research"
version = "0.1.0"
description = "Research workflow with Parallel Task MCP and Hokusai-style PDF reports"
requires-python = ">=3.11"
dependencies = [
    "weasyprint>=64.0",
    "PyHamcrest>=2.1.0",
    "markdown>=3.10",
    "httpx>=0.28.0",
    "parallel-web>=0.3.4",
    "google-genai>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.9.0",
]

[project.scripts]
research = "src.main:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
log_cli = false
log_cli_level = "CRITICAL"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
